♻️ Componentize Main part
