+++ src/pages/about.tsx
@@ -1,12 +1,5 @@
 import Head from 'next/head'
-import { Inter } from 'next/font/google'
-import styles from '@/styles/Home.module.css'
-import { Logo } from '@/components/Logo'
-import { Links } from '@/components/Links'
-import { MainImage } from '@/components/MainImage'
-import { Path } from '@/components/Path'
-
-const inter = Inter({ subsets: ['latin'] })
+import { Main } from '@/components/Main'
 
 export default function Home() {
   return (
@@ -17,14 +10,7 @@ export default function Home() {
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      <main className={`${styles.main} ${inter.className}`}>
-        <div className={styles.description}>
-          <Path page="about"/>
-          <Logo />
-        </div>
-        <MainImage />
-        <Links />
-      </main>
+      <Main page="about"/>
     </>
   )
 }
